update io toolset, more quality of life changes, update docker resources for local deployment

+++ stopDev.sh
@@ -9,16 +9,86 @@ if [ "$removeServices" == "yes" ]
 then
   echo "removing services and their underlying containers..."
 
-  docker compose -f docker-compose.mongo.yml down
-  docker compose -f docker-compose.data.yml down
-  docker compose -f docker-compose.service.yml down
+  echo "remove service layer?: (yes or no)"
+  read services
+
+  if [ "$services" == "yes" ]
+  then
+    docker compose -f docker-compose.service.yml down
+  elif [ "$services" == "no" ]
+  then
+    echo "service layer not selected for removal"
+  else
+    echo "invalid input for service layer"
+  fi
+
+  echo "remove data layer?: (yes or no)"
+  read data
+
+  if [ "$data" == "yes" ]
+  then
+    docker compose -f docker-compose.data.yml down
+  elif [ "$data" == "no" ]
+  then
+    echo "data layer not selected for removal"
+  else
+    echo "invalid input for data layer"
+  fi
+
+  echo "remove database layer?: (yes or no)"
+  read database
+
+  if [ "$database" == "yes" ]
+  then
+    docker compose -f docker-compose.db.yml down
+  elif [ "$database" == "no" ]
+  then
+    echo "database not selected for removal"
+  else
+    echo "invalid input for database"
+  fi
 elif [ "$removeServices" == "no" ]
 then
   echo "stopping services...can be restarted"
   
-  docker compose -f docker-compose.mongoreplica.yml stop
-  docker compose -f docker-compose.data.yml stop
-  docker compose -f docker-compose.service.yml stop
+  echo "stop service layer?: (yes or no)"
+  read services
+
+  if [ "$services" == "yes" ]
+  then
+    docker compose -f docker-compose.service.yml stop
+  elif [ "$services" == "no" ]
+  then
+    echo "service layer not selected for stop"
+  else
+    echo "invalid input for service layer"
+  fi
+
+  echo "stop data layer?: (yes or no)"
+  read data
+
+  if [ "$data" == "yes" ]
+  then
+    docker compose -f docker-compose.data.yml stop
+  elif [ "$data" == "no" ]
+  then
+    echo "data layer not selected for stop"
+  else
+    echo "invalid input for data layer"
+  fi
+
+  echo "stop database layer?: (yes or no)"
+  read database
+
+  if [ "$database" == "yes" ]
+  then
+    docker compose -f docker-compose.db.yml stop
+  elif [ "$database" == "no" ]
+  then
+    echo "database not selected for stop"
+  else
+    echo "invalid input for database"
+  fi
 else
   echo truthyInput
 fi
